chore(v1): commit pending V1 fixes from earlier session (calendar navigation, avatar picker, daily journey card, backup schema versioning) previously applied but never tracked

+++ lib/services/backup_service.dart
@@ -723,7 +723,7 @@ class BackupService {
         durationSeconds: (s['durationSeconds'] as num?)?.toInt(),
       );
       final newKey = await sesBox.add(obj);
-      sesKeyMap[oldId] = newKey as int;
+      sesKeyMap[oldId] = newKey;
     }
 
     // 7. Session sets
